refactor(payment): PAYPAL-3600 removed braintree connect implementation for all braintree strategies
