+++ packages/braintree-integration/src/braintree-fastlane/braintree-fastlane-payment-initialize-options.ts
@@ -17,6 +17,7 @@ import { CardInstrument } from '@bigcommerce/checkout-sdk/payment-integration-ap
  *     methodId: 'braintreeacceleratedcheckout',
  *     braintreefastlane: {
  *         onInit: (renderPayPalComponentMethod) => renderPayPalComponentMethod('#container-id'),
+ *         onChange: (showPayPalCardSelector) => showPayPalCardSelector(),
  *         styles: {
  *              root: {
  *                  backgroundColorPrimary: 'transparent',
@@ -51,9 +52,10 @@ import { CardInstrument } from '@bigcommerce/checkout-sdk/payment-integration-ap
 export default interface BraintreeFastlanePaymentInitializeOptions {
     /**
      * Is a callback that takes the CSS selector of a container
-     * where the PayPal Connect form should be inserted into.
+     * where the Braintree Fastlane form should be inserted into.
      */
     onInit?: (renderPayPalComponentMethod: (container: string) => void) => void;
+
     /**
      * Is a callback that shows Braintree stored instruments
      * when get triggered
@@ -61,9 +63,9 @@ export default interface BraintreeFastlanePaymentInitializeOptions {
     onChange?: (showPayPalCardSelector: () => Promise<CardInstrument | undefined>) => void;
 
     /**
-     * Is a stylisation options for customizing PayPal Connect components
+     * Is a stylisation options for customizing Braintree Fastlane components
      *
-     * Note: the styles for all Braintree Accelerated Checkout strategies should be the same,
+     * Note: the styles for all Braintree Fastlane strategies should be the same,
      * because they will be provided to PayPal library only for the first strategy initialization
      * no matter what strategy was initialised first
      */
