+++ packages/braintree-utils/src/braintree.ts
@@ -8,15 +8,14 @@ import {
     PaypalStyleOptions,
 } from './paypal';
 import {
-    BRAINTREE_SDK_ALPHA_VERSION,
     BRAINTREE_SDK_FASTLANE_COMPATIBLE_VERSION,
     BRAINTREE_SDK_STABLE_VERSION,
 } from './sdk-verison';
 import {
-    BraintreeClient,
     BraintreeClientCreator,
     BraintreeDataCollectorCreator,
     BraintreeError,
+    BraintreeFastlaneCreator,
     BraintreeHostedFieldsTokenizePayload,
     BraintreeModule,
     BraintreeModuleCreator,
@@ -49,13 +48,11 @@ export enum BraintreeEnv {
  */
 export interface BraintreeIntegrityValues {
     [BRAINTREE_SDK_STABLE_VERSION]?: string;
-    [BRAINTREE_SDK_ALPHA_VERSION]?: string;
     [BRAINTREE_SDK_FASTLANE_COMPATIBLE_VERSION]?: string;
 }
 
 export enum BraintreeModuleName {
     Client = 'client',
-    Connect = 'connect',
     DataCollector = 'dataCollector',
     Fastlane = 'fastlane',
     GooglePayment = 'googlePayment',
@@ -71,7 +68,6 @@ export enum BraintreeModuleName {
 
 export type BraintreeModuleCreators =
     | BraintreeClientCreator
-    | BraintreeConnectCreator
     | BraintreeFastlaneCreator
     | BraintreeDataCollectorCreator
     | GooglePayCreator
@@ -86,7 +82,6 @@ export type BraintreeModuleCreators =
 
 export interface BraintreeSDK {
     [BraintreeModuleName.Client]?: BraintreeClientCreator;
-    [BraintreeModuleName.Connect]?: BraintreeConnectCreator;
     [BraintreeModuleName.Fastlane]?: BraintreeFastlaneCreator;
     [BraintreeModuleName.DataCollector]?: BraintreeDataCollectorCreator;
     [BraintreeModuleName.GooglePayment]?: GooglePayCreator;
@@ -568,441 +563,6 @@ export interface BraintreeVisaCheckout extends BraintreeModule {
 
 /**
  *
- * Braintree Connect
- *
- */
-export type BraintreeConnectCreator = BraintreeModuleCreator<
-    BraintreeConnect,
-    BraintreeConnectConfig
->;
-
-export interface BraintreeConnectConfig {
-    authorization: string;
-    client: BraintreeClient;
-    deviceData?: string;
-    styles?: BraintreeConnectStylesOption;
-}
-
-export interface BraintreeConnect {
-    profile: BraintreeConnectProfile;
-    identity: BraintreeConnectIdentity;
-    ConnectCardComponent: (
-        options: BraintreeConnectCardComponentOptions,
-    ) => BraintreeConnectCardComponent;
-    events: BraintreeFastlaneEvents;
-}
-
-export interface BraintreeConnectProfile {
-    showCardSelector(): Promise<BraintreeFastlaneCardSelectorResponse>;
-}
-
-export interface BraintreeConnectWindow extends Window {
-    braintreeConnect: BraintreeConnect;
-}
-
-export interface BraintreeConnectIdentity {
-    lookupCustomerByEmail(email: string): Promise<BraintreeConnectLookupCustomerByEmailResult>;
-    triggerAuthenticationFlow(
-        customerId: string,
-        options?: BraintreeConnectAuthenticationOptions,
-    ): Promise<BraintreeConnectAuthenticationCustomerResult>;
-}
-
-export interface BraintreeConnectLookupCustomerByEmailResult {
-    customerContextId?: string;
-}
-
-export interface BraintreeConnectAuthenticationOptions {
-    styles?: BraintreeConnectStylesOption;
-    fetchFullProfileData?: boolean; // default: true
-}
-
-export interface BraintreeConnectStylesOption {
-    root?: {
-        backgroundColorPrimary?: string;
-        errorColor?: string;
-        fontFamily?: string;
-    };
-    input?: {
-        borderRadius?: string;
-        borderColor?: string;
-        focusBorderColor?: string;
-    };
-    toggle?: {
-        colorPrimary?: string;
-        colorSecondary?: string;
-    };
-    text?: {
-        body?: {
-            color?: string;
-            fontSize?: string;
-        };
-        caption?: {
-            color?: string;
-            fontSize?: string;
-        };
-    };
-    branding?: string; // 'light' | 'dark'
-}
-
-export interface BraintreeConnectAuthenticationCustomerResult {
-    authenticationState: BraintreeFastlaneAuthenticationState;
-    profileData: BraintreeConnectProfileData;
-}
-
-export interface BraintreeConnectProfileData {
-    connectCustomerAuthAssertionToken: string;
-    connectCustomerId: string;
-    addresses: BraintreeConnectAddress[];
-    cards: BraintreeConnectVaultedInstrument[];
-    phones: BraintreeConnectPhone[];
-    name: BraintreeConnectName;
-}
-
-export interface BraintreeConnectName {
-    given_name: string;
-    surname: string;
-}
-
-export interface BraintreeConnectAddress {
-    id?: string;
-    firstName?: string;
-    lastName?: string;
-    company?: string;
-    streetAddress: string;
-    extendedAddress?: string;
-    locality: string;
-    region: string;
-    postalCode: string;
-    countryCodeNumeric?: number;
-    countryCodeAlpha2: string;
-    countryCodeAlpha3?: string;
-    phoneNumber?: string;
-}
-
-export interface BraintreeConnectCardPaymentSource {
-    brand: string;
-    expiry: string;
-    lastDigits: string;
-    name?: string;
-    billingAddress: BraintreeConnectAddress;
-}
-
-export interface BraintreeConnectPaymentSource {
-    card: BraintreeConnectCardPaymentSource;
-}
-
-export interface BraintreeConnectVaultedInstrument {
-    id: string; // This is the nonce / token
-    paymentSource: BraintreeConnectPaymentSource;
-}
-
-export interface BraintreeConnectCardComponentOptions {
-    fields: BraintreeConnectCardComponentFields;
-}
-
-export interface BraintreeConnectCardComponentFields {
-    [key: string]: BraintreeConnectCardComponentField;
-}
-export interface BraintreeConnectCardComponentField {
-    placeholder?: string;
-    prefill?: string;
-}
-
-export interface BraintreeConnectPhone {
-    country_code: string;
-    national_number: string;
-}
-
-export interface BraintreeConnectTokenizeResult {
-    nonce: string;
-    details: BraintreeConnectTokenizeDetails;
-    description: string;
-    type: string;
-}
-
-export interface BraintreeConnectTokenizeDetails {
-    bin: string;
-    cardType: string;
-    expirationMoth: string;
-    expirationYear: string;
-    cardholderName: string;
-    lastFour: string;
-    lastTwo: string;
-}
-
-export interface BraintreeConnectTokenizeOptions {
-    billingAddress?: BraintreeConnectAddress;
-    shippingAddress?: BraintreeConnectAddress;
-}
-
-export interface Card {
-    type: string;
-    niceType: string;
-    code: {
-        name: 'CVV' | 'CID' | 'CVC';
-        size: number;
-    };
-}
-
-export interface BraintreeConnectCardComponent {
-    (options: BraintreeConnectCardComponentOptions): BraintreeConnectCardComponent;
-    tokenize(options: BraintreeConnectTokenizeOptions): Promise<BraintreeConnectTokenizeResult>;
-    getPaymentToken(
-        options: BraintreeFastlaneTokenizeOptions,
-    ): Promise<BraintreeFastlaneVaultedInstrument>; // FIXME: this method does not support by Connect
-    render(element: string): void;
-}
-
-/**
- *  Braintree Fastlane
- */
-export type BraintreeFastlaneCreator = BraintreeModuleCreator<
-    BraintreeFastlane,
-    BraintreeFastlaneConfig
->;
-
-export interface BraintreeFastlaneConfig {
-    authorization: string;
-    client: BraintreeClient;
-    deviceData?: string;
-    styles?: BraintreeFastlaneStylesOption;
-}
-
-export interface BraintreeFastlane {
-    identity: BraintreeFastlaneIdentity;
-    profile: BraintreeFastlaneProfile;
-    FastlaneCardComponent: (
-        options: BraintreeFastlaneCardComponentOptions,
-    ) => Promise<BraintreeFastlaneCardComponent>;
-    events: BraintreeFastlaneEvents;
-}
-
-export interface BraintreeFastlaneProfile {
-    showCardSelector(): Promise<BraintreeFastlaneCardSelectorResponse>;
-    showShippingAddressSelector(): Promise<BraintreeFastlaneShippingAddressSelectorResponse>;
-}
-
-export interface BraintreeFastlaneShippingAddressSelectorResponse {
-    selectionChanged: boolean;
-    selectedAddress: BraintreeFastlaneShippingAddress;
-}
-
-export interface BraintreeFastlaneShippingAddress {
-    name: BraintreeFastlaneProfileName;
-    phoneNumber: string;
-    id?: string;
-    firstName?: string;
-    lastName?: string;
-    company?: string;
-    streetAddress: string;
-    extendedAddress?: string;
-    locality: string;
-    region: string;
-    postalCode: string;
-    countryCodeNumeric?: number;
-    countryCodeAlpha2: string;
-    countryCodeAlpha3?: string;
-}
-
-export interface BraintreeFastlaneProfileName {
-    fullName: string;
-    firstName?: string;
-    lastName?: string;
-}
-
-export interface BraintreeFastlaneCardSelectorResponse {
-    selectionChanged: boolean;
-    selectedCard: BraintreeFastlaneVaultedInstrument;
-}
-
-export interface BraintreeFastlaneWindow extends Window {
-    braintreeFastlane: BraintreeFastlane;
-}
-
-export interface BraintreeFastlaneIdentity {
-    lookupCustomerByEmail(email: string): Promise<BraintreeFastlaneLookupCustomerByEmailResult>;
-    triggerAuthenticationFlow(
-        customerId: string,
-        options?: BraintreeFastlaneAuthenticationOptions,
-    ): Promise<BraintreeFastlaneAuthenticationCustomerResult>;
-}
-
-export interface BraintreeFastlaneLookupCustomerByEmailResult {
-    customerContextId?: string;
-}
-
-export interface BraintreeFastlaneAuthenticationOptions {
-    styles?: BraintreeFastlaneStylesOption;
-}
-
-export interface BraintreeFastlaneStylesOption {
-    root?: {
-        backgroundColorPrimary?: string;
-        errorColor?: string;
-        fontFamily?: string;
-    };
-    input?: {
-        borderRadius?: string;
-        borderColor?: string;
-        focusBorderColor?: string;
-    };
-    toggle?: {
-        colorPrimary?: string;
-        colorSecondary?: string;
-    };
-    text?: {
-        body?: {
-            color?: string;
-            fontSize?: string;
-        };
-        caption?: {
-            color?: string;
-            fontSize?: string;
-        };
-    };
-    branding?: string; // 'light' | 'dark'
-}
-
-export enum BraintreeFastlaneAuthenticationState {
-    SUCCEEDED = 'succeeded',
-    FAILED = 'failed',
-    CANCELED = 'cancelled',
-    UNRECOGNIZED = 'unrecognized',
-}
-
-export interface BraintreeFastlaneAuthenticationCustomerResult {
-    authenticationState: BraintreeFastlaneAuthenticationState;
-    profileData: BraintreeFastlaneProfileData;
-}
-
-export interface BraintreeFastlaneProfileData {
-    fastlaneCustomerAuthAssertionToken: string;
-    fastlaneCustomerId: string;
-    shippingAddress: BraintreeFastlaneAddress;
-    card: BraintreeFastlaneVaultedInstrument;
-    name: BraintreeFastlaneName;
-}
-
-export interface BraintreeFastlaneName {
-    firstName: string;
-    lastName: string;
-}
-
-export interface BraintreeFastlaneAddress {
-    id?: string;
-    firstName?: string;
-    lastName?: string;
-    company?: string;
-    streetAddress: string;
-    extendedAddress?: string;
-    locality: string;
-    region: string;
-    postalCode: string;
-    countryCodeNumeric?: number;
-    countryCodeAlpha2: string;
-    countryCodeAlpha3?: string;
-    phoneNumber?: string;
-}
-
-export interface BraintreeFastlaneCardPaymentSource {
-    brand: string;
-    expiry: string;
-    lastDigits: string;
-    name?: string;
-    billingAddress: BraintreeFastlaneAddress;
-}
-
-export interface BraintreeFastlanePaymentSource {
-    card: BraintreeFastlaneCardPaymentSource;
-}
-
-export interface BraintreeFastlaneVaultedInstrument {
-    id: string; // This is the nonce / token
-    paymentSource: BraintreeFastlanePaymentSource;
-}
-
-export interface BraintreeFastlaneCardComponentOptions {
-    styles: BraintreeFastlaneStylesOption;
-    fields: BraintreeFastlaneCardComponentFields;
-}
-
-export interface BraintreeFastlaneCardComponentFields {
-    cardholderName?: {
-        enabled?: boolean;
-        prefill?: string;
-    };
-    phoneNumber?: {
-        placeholder?: string;
-        prefill?: string;
-    };
-}
-
-export interface BraintreeFastlaneTokenizeDetails {
-    bin: string;
-    cardType: string;
-    expirationMoth: string;
-    expirationYear: string;
-    cardholderName: string;
-    lastFour: string;
-    lastTwo: string;
-}
-
-export interface BraintreeFastlaneTokenizeOptions {
-    billingAddress?: BraintreeFastlaneAddress;
-    shippingAddress?: BraintreeFastlaneAddress;
-}
-
-export interface BraintreeFastlaneEvents {
-    apmSelected: (options: BraintreeFastlaneApmSelectedEventOptions) => void;
-    emailSubmitted: (options: BraintreeFastlaneEmailEnteredEventOptions) => void;
-    orderPlaced: (options: BraintreeFastlaneOrderPlacedEventOptions) => void;
-}
-
-export interface BraintreeFastlaneEventCommonOptions {
-    context_type: 'cs_id';
-    context_id: string; // checkout session id
-    page_type: 'checkout_page';
-    page_name: string; // title of the checkout initiation page
-    partner_name: 'bigc';
-    user_type: 'store_member' | 'store_guest'; // type of the user on the merchant site
-    store_id: string;
-    merchant_name: string;
-    experiment: string; // stringify JSON object "[{ treatment_group: 'test' | 'control' }]"
-}
-
-export interface BraintreeFastlaneApmSelectedEventOptions
-    extends BraintreeFastlaneEventCommonOptions {
-    apm_shown: '0' | '1'; // alternate payment shown on the checkout page
-    apm_list: string; // list of alternate payment shown on checkout page
-    apm_selected: string; // alternate payment method selected / methodId
-    apm_location: 'pre-email section' | 'payment section'; // placement of APM, whether it be above the email entry or in the radio buttons
-}
-
-export interface BraintreeFastlaneEmailEnteredEventOptions
-    extends BraintreeFastlaneEventCommonOptions {
-    user_email_saved: boolean; // shows whether checkout was loaded with or without a saved email
-    apm_shown: '0' | '1'; // alternate payment shown on the checkout page
-    apm_list: string; // list of alternate payment shown on checkout page 'applepay,googlepay,paypal'
-}
-
-export interface BraintreeFastlaneOrderPlacedEventOptions
-    extends BraintreeFastlaneEventCommonOptions {
-    selected_payment_method: string;
-    currency_code: string;
-}
-
-export interface BraintreeFastlaneCardComponent {
-    (options: BraintreeFastlaneCardComponentOptions): BraintreeFastlaneCardComponent;
-    tokenize(options: BraintreeConnectTokenizeOptions): Promise<BraintreeConnectTokenizeResult>; // FIXME: this method does not support by Fastlane
-    getPaymentToken(
-        options: BraintreeFastlaneTokenizeOptions,
-    ): Promise<BraintreeFastlaneVaultedInstrument>;
-    render(element: string): void;
-}
-
-/**
- *
  * Braintree Local Methods
  *
  */
@@ -1136,7 +696,5 @@ export interface GooglePayBraintreeSDK extends BraintreeModule {
 export interface BraintreeHostWindow extends BraintreeWindow {
     braintree?: BraintreeSDK;
     paypal?: PaypalSDK;
-    braintreeConnect?: BraintreeConnect;
-    braintreeFastlane?: BraintreeFastlane;
     V?: VisaCheckoutSDK;
 }
