+++ packages/braintree-utils/src/sdk-verison.ts
@@ -1,3 +1,2 @@
 export const BRAINTREE_SDK_STABLE_VERSION = '3.95.0';
-export const BRAINTREE_SDK_ALPHA_VERSION = '3.95.0-connect-alpha.12';
 export const BRAINTREE_SDK_FASTLANE_COMPATIBLE_VERSION = '3.101.0-fastlane-beta.7.2';
