+++ packages/braintree-utils/src/utils/index.ts
@@ -1,8 +1,3 @@
 export { default as isBraintreeAcceleratedCheckoutCustomer } from './is-braintree-accelerated-checkout-customer';
-export { default as isBraintreeConnectWindow } from './is-braintree-connect-window';
 export { default as isBraintreeError } from './is-braintree-error';
 export { default as isBraintreeFastlaneWindow } from './is-braintree-fastlane-window';
-export { default as isBraintreeFastlaneProfileData } from './is-braintree-fastlane-profile-data';
-export { default as isBraintreeConnectProfileData } from './is-braintree-connect-profile-data';
-export { default as isBraintreeConnectName } from './is-braintree-connect-name';
-export { default as isBraintreeConnectPhone } from './is-braintree-connect-phone';
