--- packages/braintree-utils/src/utils/is-braintree-connect-phone.ts
@@ -1,14 +1,0 @@
-import { BraintreeConnectPhone } from '../';
-
-export default function isBraintreeConnectPhone(
-    braintreeConnectPhone: BraintreeConnectPhone | string | undefined,
-): braintreeConnectPhone is BraintreeConnectPhone {
-    if (!braintreeConnectPhone) {
-        return false;
-    }
-
-    return (
-        braintreeConnectPhone.hasOwnProperty('country_code') &&
-        braintreeConnectPhone.hasOwnProperty('national_number')
-    );
-}
