--- packages/braintree-utils/src/utils/is-braintree-connect-window.ts
@@ -1,5 +1,0 @@
-import { BraintreeConnectWindow } from '../braintree';
-
-export default function isBraintreeConnectWindow(window: Window): window is BraintreeConnectWindow {
-    return Boolean(window.hasOwnProperty('braintreeConnect'));
-}
