--- packages/braintree-utils/src/utils/is-braintree-fastlane-profile-data.ts
@@ -1,17 +1,0 @@
-import { BraintreeConnectProfileData, BraintreeFastlaneProfileData } from '../';
-
-export default function isBraintreeFastlaneProfileData(
-    profileData: BraintreeFastlaneProfileData | BraintreeConnectProfileData | undefined,
-): profileData is BraintreeFastlaneProfileData {
-    if (!profileData) {
-        return false;
-    }
-
-    return (
-        profileData.hasOwnProperty('shippingAddress') &&
-        profileData.hasOwnProperty('card') &&
-        profileData.hasOwnProperty('fastlaneCustomerId') &&
-        profileData.hasOwnProperty('fastlaneCustomerAuthAssertionToken') &&
-        profileData.hasOwnProperty('name')
-    );
-}
